Create skeleton file and renames files on molecules folder
