+++ app/ui/molecules/mol-table-save-form.tsx
@@ -1,5 +1,5 @@
 'use client'
-import { printAction, SaveActionMOCK } from "@/app/(routes)/table/actions"
+import { printAction, SaveActionMOCK } from "@/app/(routes)/table/create/actions"
 import TablePanel from "./mol-table-panel"
 import { SubmitButton } from "../atoms/atom-button"
 import { Action_State } from "@/app/lib/definitions"
